--- a/android/src/main/java/expo/modules/xmtpreactnativesdk/wrappers/DmWrapper.kt
+++ b/android/src/main/java/expo/modules/xmtpreactnativesdk/wrappers/DmWrapper.kt
@@ -21,7 +21,6 @@ class DmWrapper {
                 put("version", "DM")
                 put("topic", dm.topic)
                 put("peerInboxId", dm.peerInboxId)
-                if (dmParams.creatorInboxId) put("creatorInboxId", dm.creatorInboxId())
                 if (dmParams.consentState) {
                     put("consentState", consentStateToString(dm.consentState()))
                 }