--- a/android/src/main/java/expo/modules/xmtpreactnativesdk/wrappers/GroupWrapper.kt
+++ b/android/src/main/java/expo/modules/xmtpreactnativesdk/wrappers/GroupWrapper.kt
@@ -20,9 +20,6 @@ class GroupWrapper {
                 put("createdAt", group.createdAt.time)
                 put("version", "GROUP")
                 put("topic", group.topic)
-                if (groupParams.members) {
-                    put("members", group.members().map { MemberWrapper.encode(it) })
-                }
                 if (groupParams.creatorInboxId) put("creatorInboxId", group.creatorInboxId())
                 if (groupParams.isActive) put("isActive", group.isActive())
                 if (groupParams.addedByInboxId) put("addedByInboxId", group.addedByInboxId())
@@ -54,7 +51,6 @@ class GroupWrapper {
 }
 
 class ConversationParamsWrapper(
-    val members: Boolean = true,
     val creatorInboxId: Boolean = true,
     val isActive: Boolean = true,
     val addedByInboxId: Boolean = true,
@@ -69,7 +65,6 @@ class ConversationParamsWrapper(
             if (conversationParams.isEmpty()) return ConversationParamsWrapper()
             val jsonOptions = JsonParser.parseString(conversationParams).asJsonObject
             return ConversationParamsWrapper(
-                if (jsonOptions.has("members")) jsonOptions.get("members").asBoolean else true,
                 if (jsonOptions.has("creatorInboxId")) jsonOptions.get("creatorInboxId").asBoolean else true,
                 if (jsonOptions.has("isActive")) jsonOptions.get("isActive").asBoolean else true,
                 if (jsonOptions.has("addedByInboxId")) jsonOptions.get("addedByInboxId").asBoolean else true,