--- a/src/index.ts
+++ b/src/index.ts
@@ -298,11 +298,7 @@ export async function findOrCreateDm<
   const dm = JSON.parse(
     await XMTPModule.findOrCreateDm(client.inboxId, peerAddress)
   )
-  const members =
-    dm['members']?.map((mem: string) => {
-      return Member.from(mem)
-    }) || []
-  return new Dm(client, dm, members)
+  return new Dm(client, dm)
 }
 
 export async function createGroup<
@@ -331,11 +327,7 @@ export async function createGroup<
     )
   )
 
-  const members =
-    group['members']?.map((mem: string) => {
-      return Member.from(mem)
-    }) || []
-  return new Group(client, group, members)
+  return new Group(client, group)
 }
 
 export async function createGroupCustomPermissions<
@@ -363,11 +355,8 @@ export async function createGroupCustomPermissions<
       JSON.stringify(options)
     )
   )
-  const members =
-    group['members']?.map((mem: string) => {
-      return Member.from(mem)
-    }) || []
-  return new Group(client, group, members)
+
+  return new Group(client, group)
 }
 
 export async function listGroups<
@@ -387,14 +376,11 @@ export async function listGroups<
     )
   ).map((json: string) => {
     const group = JSON.parse(json)
-    const members =
-      group['members']?.map((mem: string) => {
-        return Member.from(mem)
-      }) || []
+
     const lastMessage = group['lastMessage']
       ? DecodedMessage.from(group['lastMessage'], client)
       : undefined
-    return new Group(client, group, members, lastMessage)
+    return new Group(client, group, lastMessage)
   })
 }
 
@@ -415,14 +401,15 @@ export async function listV3Conversations<
     )
   ).map((json: string) => {
     const jsonObj = JSON.parse(json)
-    const members =
-      jsonObj.members?.map((mem: string) => {
-        return Member.from(mem)
-      }) || []
+
+    const lastMessage = jsonObj['lastMessage']
+      ? DecodedMessage.from(jsonObj['lastMessage'], client)
+      : undefined
+
     if (jsonObj.version === ConversationVersion.GROUP) {
-      return new Group(client, jsonObj, members)
+      return new Group(client, jsonObj, lastMessage)
     } else {
-      return new Dm(client, jsonObj, members)
+      return new Dm(client, jsonObj, lastMessage)
     }
   })
 }
@@ -520,11 +507,8 @@ export async function findGroup<
   if (!group || Object.keys(group).length === 0) {
     return undefined
   }
-  const members =
-    group['members']?.map((mem: string) => {
-      return Member.from(mem)
-    }) || []
-  return new Group(client, group, members)
+
+  return new Group(client, group)
 }
 
 export async function findConversation<
@@ -538,15 +522,11 @@ export async function findConversation<
   if (!conversation || Object.keys(conversation).length === 0) {
     return undefined
   }
-  const members =
-    conversation['members']?.map((mem: string) => {
-      return Member.from(mem)
-    }) || []
 
   if (conversation.version === ConversationVersion.GROUP) {
-    return new Group(client, conversation, members)
+    return new Group(client, conversation)
   } else {
-    return new Dm(client, conversation, members)
+    return new Dm(client, conversation)
   }
 }
 
@@ -561,15 +541,11 @@ export async function findConversationByTopic<
   if (!conversation || Object.keys(conversation).length === 0) {
     return undefined
   }
-  const members =
-    conversation['members']?.map((mem: string) => {
-      return Member.from(mem)
-    }) || []
 
   if (conversation.version === ConversationVersion.GROUP) {
-    return new Group(client, conversation, members)
+    return new Group(client, conversation)
   } else {
-    return new Dm(client, conversation, members)
+    return new Dm(client, conversation)
   }
 }
 
@@ -584,11 +560,8 @@ export async function findDm<
   if (!dm || Object.keys(dm).length === 0) {
     return undefined
   }
-  const members =
-    dm['members']?.map((mem: string) => {
-      return Member.from(mem)
-    }) || []
-  return new Dm(client, dm, members)
+
+  return new Dm(client, dm)
 }
 
 export async function findV3Message<
@@ -849,11 +822,7 @@ export async function listAll<
   return list.map((json: string) => {
     const jsonObj = JSON.parse(json)
     if (jsonObj.version === ConversationVersion.GROUP) {
-      const members =
-        jsonObj.members?.map((mem: string) => {
-          return Member.from(mem)
-        }) || []
-      return new Group(client, jsonObj, members)
+      return new Group(client, jsonObj)
     } else {
       return new Conversation(client, jsonObj)
     }
@@ -1471,11 +1440,7 @@ export async function processWelcomeMessage<
     encryptedMessage
   )
   const group = JSON.parse(json)
-  const members =
-    group['members']?.map((mem: string) => {
-      return Member.from(mem)
-    }) || []
-  return new Group(client, group, members)
+  return new Group(client, group)
 }
 
 export async function processConversationWelcomeMessage<
@@ -1489,15 +1454,11 @@ export async function processConversationWelcomeMessage<
     encryptedMessage
   )
   const conversation = JSON.parse(json)
-  const members =
-    conversation['members']?.map((mem: string) => {
-      return Member.from(mem)
-    }) || []
 
   if (conversation.version === ConversationVersion.GROUP) {
-    return new Group(client, conversation, members)
+    return new Group(client, conversation)
   } else {
-    return new Dm(client, conversation, members)
+    return new Dm(client, conversation)
   }
 }