--- a/src/lib/Dm.ts
+++ b/src/lib/Dm.ts
@@ -15,7 +15,6 @@ import * as XMTP from '../index'
 export interface DmParams {
   id: string
   createdAt: number
-  creatorInboxId: InboxId
   topic: string
   consentState: ConsentState
   lastMessage?: DecodedMessage