--- a/src/lib/types/GroupOptions.ts
+++ b/src/lib/types/GroupOptions.ts
@@ -1,5 +1,4 @@
 export type GroupOptions = {
-  members?: boolean
   creatorInboxId?: boolean
   isActive?: boolean
   addedByInboxId?: boolean