Remove members and creatorInboxId from the listing of groups because it will never be performant.

Also upgrades to the version of `peerInboxId` that is more performant.

Android still is a good bit slower.

Here is

**iOS**

| Metric                   | V2Alix          | V3Davon         |
|--------------------------|-----------------|-----------------|
| 50 DMs (1st load)        | 116 ms         | 275 ms         |
| 50 DMs (2nd load)        | 25 ms          | 9 ms           |
| 55 DMs (1st load)        | 19 ms          | 34 ms          |
| 55 DMs (2nd load)        | 9 ms           | 9 ms           |

**Android**

| Metric                   | V2Alix          | V3Davon         |
|--------------------------|-----------------|-----------------|
| 50 DMs (1st load)        | 364 ms         | 538 ms         |
| 50 DMs (2nd load)        | 55 ms          | 62 ms          |
| 55 DMs (1st load)        | 56 ms          | 61 ms          |
| 55 DMs (2nd load)        | 25 ms          | 62 ms          |